fix: enhance Assert with IsNull/IsNotNull + numeric AreEqual overloads; fix 5 Text.Json C# bugs
- Add IsNull, IsNotNull to Assert class (Chaos.TestFramework.Sdk)
- Add 12 numeric AreEqual overloads (long, decimal, double, float, uint, ulong,
  byte, short, ushort, sbyte, char, bool) to support all types used in subjects
- Fix serializer: refactor CS8175 (ref struct Utf8JsonReader captured in lambda)
  by moving reader construction inside lambda body
- Fix reader: cast JsonTokenType enum to int for Assert.AreEqual; use 0L for
  long-typed TokenStartIndex assertion
- Fix document-element: cast JsonValueKind enum to int for Assert.AreEqual
- Fix polymorphism-reference: replace invalid ReferenceHandler<int> (doesn't
  satisfy ReferenceResolver constraint) with Preserve-based tests
- Fix metadata-typeinfo: add 'using Chaos.TestFramework;' + replace read-only
  property assignments with reflection SetValue to compile on .NET 10

+++ testing/foundation-dll/System.Text.Json/document-element/managed/subjects/DocumentElementSubjects.cs
@@ -31,7 +31,7 @@ public static partial class DocumentElementSubjects
     // [1] System.Text.Json/System.Text.Json.JsonDocument::get_RootElement:System.Text.Json.JsonElement()
     public static void CustomEntrySubject_1()
     {
-        try { using var doc = JsonDocument.Parse(TestJson); var el = doc.RootElement; Assert.AreEqual(JsonValueKind.Object, el.ValueKind); }
+        try { using var doc = JsonDocument.Parse(TestJson); var el = doc.RootElement; Assert.AreEqual((int)JsonValueKind.Object, (int)el.ValueKind); }
         catch { _exitCode = 1; }
     }
 
@@ -136,7 +136,7 @@ public static partial class DocumentElementSubjects
     // [16] System.Text.Json/System.Text.Json.JsonElement::get_ValueKind:System.Text.Json.JsonValueKind()
     public static void CustomEntrySubject_16()
     {
-        try { using var doc = JsonDocument.Parse(TestJson); Assert.AreEqual(JsonValueKind.Object, doc.RootElement.ValueKind); }
+        try { using var doc = JsonDocument.Parse(TestJson); Assert.AreEqual((int)JsonValueKind.Object, (int)doc.RootElement.ValueKind); }
         catch { _exitCode = 1; }
     }
